Refactor: Update component references from DTX.UI.Components to DTX.Song.Components
- Changed all instances of DTX.UI.Components to DTX.Song.Components in the following files:
  - SongSelectionStage.cs
  - DefaultGraphicsGenerator.cs
  - GraphicsExtensionsTests.cs
  - SongNavigationPerformanceTests.cs
  - CurvedLayoutTests.cs
  - SongBarRendererTests.cs
  - SongBarTests.cs
  - SongListDisplayTests.cs
  - SongStatusPanelTests.cs
  - UIArchitectureTests.cs
