+++ DTXMania.Test/Performance/SongNavigationPerformanceTests.cs
@@ -4,7 +4,7 @@ using System.Diagnostics;
 using System.Linq;
 using Xunit;
 using DTX.Song;
-using DTX.UI.Components;
+using DTX.Song.Components;
 using DTXMania.Game.Lib.Song.Entities;
 using SongScore = DTXMania.Game.Lib.Song.Entities.SongScore;
