+++ DTXMania.Test/UI/SongBarTests.cs
@@ -1,7 +1,7 @@
 using Xunit;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using DTX.UI.Components;
+using DTX.Song.Components;
 using DTX.Song;
 using DTX.Resources;
 using DTXMania.Test.Helpers;
