+++ DTXMania.Test/UI/UIArchitectureTests.cs
@@ -1,7 +1,7 @@
 using Moq;
 using Xunit;
 using DTX.UI;
-using DTX.UI.Components;
+using DTX.Song.Components;
 using DTX.Song;
 using Microsoft.Xna.Framework;
 using System;
